+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Windows installer for easy setup of the native messaging host setup (sets up Chrome,
+  Chrome Beta, Chrome Canary, Chromium, Firefox, and Opera if they are installed).
+  - Also bundles a copy of `mpv` and `yt-dlp`.
+- Post-installation page for browser extension with platform-specific instructions.
+- Post-uninstallation page for browser extension.
+
+### Changed
+
+- Removed Linux-specific `--gpu-api=opengl` flag on Windows.
+
 ## [0.1.3]
 
 ### Fixed